Disambiguate something in the README
Change phrasing of description of `--reruns=N` parameter in the README. Its previous phrasing referred to rerunning "failing tests N times" and the phrasing made it ambiguous whether "N" is a shared count, or a per-test count. Indeed, it reruns *each* test, *up to* N times. This commit attempts to make that completely clear.