fix: don't bind all ips
Binding all IPs is kind of Security issues and also very annoying when
the OS has a firewall blocking application doing such thing or opening a
popup to allow it (like the default MacOS or Windows firewall).

The client uses "localhost", so the server can only bind the ip
associated with the hostname "localhost".

This will avoid opening this tools to anyone on the network and avoid
having to configure windows or macos firewall to make pytest run.