Allow to pass a callable condition to the `flaky` marker (#299)
* feat: add 'condition' parameter to flaky marker

This introduces a condition parameter to the `@pytest.mark.flaky`
decorator, allowing for more granular control over when a test is rerun.

The condition can be a callable or a string.

If it's a callable, it will be passed the exception object from the
failed test. The test will be rerun only if the callable returns True.

If it's a string, it will be evaluated with the following objects in
its global context: os, sys, platform, config, and error (the
exception instance). The test will be rerun only if the string
evaluates to True.

Exceptions raised by a condition callable are now caught and logged as a
warning, preventing the test suite from crashing.

Issue #230


---------

Co-authored-by: Michael Howitz <m.howitz@minddistrict.com>